# Conflicts:
#	samples/opengl/CMakeLists.txt
#	samples/opengl/forwardPointCloudRendering/CMakeLists.txt
#	samples/opengl/forwardPointCloudRendering/forward_point_cloud_rendering.cpp
#	samples/opengl/renderPipelineSample/pipelines_sample.cpp
#	samples/opengl/sample_opengl_cascaded_shadow_maps.cpp
#	samples/opengl/sample_opengl_cornell.cpp
#	samples/opengl/simpleGLFWWindow/simple_glfw_window.cpp
#	shader/lighting/uber_lighting_helpers.glsl
#	src/saiga/core/geometry/triangle_mesh_generator.cpp
#	src/saiga/opengl/rendering/deferredRendering/deferred_renderer.h
#	src/saiga/opengl/rendering/deferredRendering/uber_deferred_renderer.cpp
#	src/saiga/opengl/rendering/deferredRendering/uber_deferred_renderer.h
#	src/saiga/opengl/rendering/forwardRendering/forward_renderer.cpp
#	src/saiga/opengl/rendering/forwardRendering/forward_renderer.h
#	src/saiga/opengl/rendering/lighting/all.h
#	src/saiga/opengl/rendering/lighting/deferred_lighting.cpp
#	src/saiga/opengl/rendering/lighting/deferred_lighting.h
#	src/saiga/opengl/rendering/lighting/directional_light.h
#	src/saiga/opengl/rendering/lighting/forward_lighting.cpp
#	src/saiga/opengl/rendering/lighting/forward_lighting.h
#	src/saiga/opengl/rendering/lighting/light.cpp
#	src/saiga/opengl/rendering/lighting/light_clusterer.cpp
#	src/saiga/opengl/rendering/lighting/light_clusterer.h
#	src/saiga/opengl/rendering/lighting/point_light.h
#	src/saiga/opengl/rendering/lighting/renderer_lighting.cpp
#	src/saiga/opengl/rendering/lighting/renderer_lighting.h
#	src/saiga/opengl/rendering/lighting/spot_light.h
#	src/saiga/opengl/rendering/lighting/uber_deferred_lighting.cpp
#	src/saiga/opengl/rendering/lighting/uber_deferred_lighting.h
#	src/saiga/opengl/window/RendererSampleWindow.h

+++ samples/opengl/depthMapTriangulation/decimate.cpp
@@ -464,11 +464,11 @@ bool QuadricDecimater::check_for_folding_triangles(const MyMesh::EdgeHandle edge
     }
 
     // check them against each other
-    for (int i = 0; i < normals.size(); ++i)
+    for (int i = 0; i < (int)normals.size(); ++i)
     {
         OpenMesh::Vec3f normal_1 = normals[i];
 
-        for (int j = i + 1; j < normals.size(); ++j)
+        for (int j = i + 1; j < (int)normals.size(); ++j)
         {
             // dot(A,B) = |A| * |B| * cos(angle)
             // which can be rearranged to
