+++ src/saiga/opengl/rendering/VRRendering/VRRenderer.h
@@ -48,7 +48,7 @@ class SAIGA_OPENGL_API VRRenderer : public OpenGLRenderer
     mat4 viewLeft, viewRight;
 
     std::shared_ptr<PostProcessingShader> framebufferToDebugWindowShader;
-    IndexedVertexBuffer<VertexNT, GLushort> quadMesh;
+    IndexedVertexBuffer<VertexNT, uint32_t> quadMesh;
 };
 
 }  // namespace Saiga
